#980 Moved console log related operations out of artifact service into it's own service
