+++ server/src/com/thoughtworks/go/domain/activity/ConsoleLogArtifactHandler.java
@@ -21,7 +21,7 @@ package com.thoughtworks.go.domain.activity;
 import com.thoughtworks.go.domain.JobIdentifier;
 import com.thoughtworks.go.domain.JobInstance;
 import com.thoughtworks.go.server.domain.JobStatusListener;
-import com.thoughtworks.go.server.service.ArtifactsService;
+import com.thoughtworks.go.server.service.ConsoleService;
 import com.thoughtworks.go.util.GoConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -30,11 +30,11 @@ import static java.lang.String.format;
 
 @Component
 public class ConsoleLogArtifactHandler implements JobStatusListener {
-    private ArtifactsService service;
+    private ConsoleService consoleService;
 
     @Autowired
-    public ConsoleLogArtifactHandler(ArtifactsService service) {
-        this.service = service;
+    public ConsoleLogArtifactHandler(ConsoleService consoleService) {
+        this.consoleService = consoleService;
     }
 
     @Override
@@ -42,9 +42,9 @@ public class ConsoleLogArtifactHandler implements JobStatusListener {
         if (job.getState().isCompleted()) {
             try {
                 JobIdentifier identifier = job.getIdentifier();
-                service.moveConsoleArtifacts(identifier);
+                consoleService.moveConsoleArtifacts(identifier);
                 // TODO: Put correct timestamp of job completion the agent and the server maybe on different timezones.
-                service.appendToConsoleLog(identifier, format("[%s] %s %s", GoConstants.PRODUCT_NAME, "Job Completed"
+                consoleService.appendToConsoleLog(identifier, format("[%s] %s %s", GoConstants.PRODUCT_NAME, "Job Completed"
                         , identifier.buildLocatorForDisplay()));
             } catch (Exception e) {
                 throw new RuntimeException(e);
